Timer audio fixes (#1033)
* Timer alert fixes

Resolved issue with timerelapsed audio not working, modified logic so it leverages rfsuite.session.timer.live to track alerting, correct some logic errors in the code. Everything tested and works as expected. Renamed the event to flighttimer.lua so it doesnt clash naming wise with the existing timer.lua event.

* Modified beep logic to be based on rfsuite.session.timer.live.

---------

Co-authored-by: Rob Thomson <rob@marotori.com>